+++ go.mod
@@ -4,6 +4,7 @@ go 1.16
 
 require (
 	github.com/SCU-SJL/menuscreen v1.0.5
+	github.com/SCU-SJL/sinfra v0.0.0-20220929071109-a4dc7f29ae01
 	github.com/gdamore/tcell/v2 v2.5.3 // indirect
 	github.com/mattn/go-runewidth v0.0.14 // indirect
 	github.com/rivo/uniseg v0.4.2 // indirect
